Rename module cache to domain.
